refactor: Remove MVC mess
