+++ src/bot/commands/feeds/mod.rs
@@ -1,15 +1,10 @@
 use crate::bot::commands::Cog;
 use crate::bot::commands::Context;
 use crate::bot::commands::Error;
-use crate::bot::commands::feeds::controller::SettingsController;
-use crate::bot::commands::feeds::controller::SubscribeController;
-use crate::bot::commands::feeds::controller::SubscriptionsController;
-use crate::bot::commands::feeds::controller::UnsubscribeController;
-use crate::bot::commands::feeds::model::SendInto;
+use crate::bot::commands::feeds::commands::SendInto;
 
-pub mod controller;
-pub mod model;
-pub mod view;
+pub mod commands;
+pub mod views;
 
 pub struct FeedsCog;
 
@@ -32,7 +27,7 @@ impl FeedsCog {
         default_member_permissions = "ADMINISTRATOR | MANAGE_GUILD"
     )]
     pub async fn settings(ctx: Context<'_>) -> Result<(), Error> {
-        SettingsController::execute(ctx).await
+        commands::settings(ctx).await
     }
 
     #[poise::command(slash_command)]
@@ -45,7 +40,7 @@ impl FeedsCog {
             SendInto,
         >,
     ) -> Result<(), Error> {
-        SubscribeController::execute(ctx, links, send_into).await
+        commands::subscribe(ctx, links, send_into).await
     }
 
     #[poise::command(slash_command)]
@@ -58,7 +53,7 @@ impl FeedsCog {
             SendInto,
         >,
     ) -> Result<(), Error> {
-        UnsubscribeController::execute(ctx, links, send_into).await
+        commands::unsubscribe(ctx, links, send_into).await
     }
 
     #[poise::command(slash_command)]
@@ -68,21 +63,21 @@ impl FeedsCog {
             SendInto,
         >,
     ) -> Result<(), Error> {
-        SubscriptionsController::execute(ctx, sent_into).await
+        commands::subscriptions(ctx, sent_into).await
     }
 
     async fn autocomplete_subscriptions<'a>(
         ctx: Context<'_>,
         partial: &str,
     ) -> poise::serenity_prelude::CreateAutocompleteResponse<'a> {
-        UnsubscribeController::autocomplete_subscriptions(ctx, partial).await
+        commands::autocomplete_subscriptions(ctx, partial).await
     }
 
     async fn autocomplete_supported_feeds<'a>(
         ctx: Context<'_>,
         partial: &str,
     ) -> poise::serenity_prelude::CreateAutocompleteResponse<'a> {
-        SubscribeController::autocomplete_supported_feeds(ctx, partial).await
+        commands::autocomplete_supported_feeds(ctx, partial).await
     }
 }
